+++ src/index.js
@@ -1,4 +1,7 @@
 import { render } from './modules/domStuff';
 import './styles.css';
+import { todobtn, formEvent } from './modules/logic';
 
+todobtn();
+formEvent();
 render();
